Разделил классы на entities и repositories.
