--- models/Basket.php
@@ -1,43 +1,0 @@
-<?php
-namespace app\models;
-
-
-use app\engine\Db;
-
-class Basket extends DbModel
-{
-    protected $id;
-    protected $session_id;
-    protected $product_id;
-    protected $user_id;
-
-    protected $props = [
-        'session_id' => false,
-        'product_id' => false,
-        'user_id' => false
-    ];
-
-    public function __construct($session_id = null, $product_id = null, $user_id = null)
-    {
-        $this->session_id = $session_id;
-        $this->product_id = $product_id;
-        $this->user_id = $user_id;
-    }
-
-    public static function getBasket($session)
-    {
-        $sql = "SELECT p.id id_prod, b.id id_basket, p.name, p.description, p.price, p.img FROM basket b, products p WHERE b.product_id=p.id AND session_id = :session";
-        return Db::getInstance()->queryAll($sql, ['session' => $session]);
-    }
-
-    public static function deleteFromBasket()
-    {
-//        return Db::getInstance()->queryAll($sql, ['session' => $session]);
-    }
-
-    public static function getTableName()
-    {
-        return "basket";
-    }
-
-}
