+++ models/Repository.php
@@ -5,7 +5,7 @@ namespace app\models;
 use app\engine\Db;
 use app\interfaces\IModel;
 
-abstract class DbModel extends Model
+abstract class Repository implements IModel
 {
     public static function getOneWhere($field, $value) {
         $tableName = static::getTableName();
@@ -18,7 +18,7 @@ abstract class DbModel extends Model
         $tableName = static::getTableName();
         $sql = "SELECT count(id) as count FROM {$tableName} WHERE `$field`=:value";
         return Db::getInstance()->queryOne($sql, ["value" => $value])['count'];
-    } //SELECT count(id) as count FROM basket WHERE session_id = "07je2u80mgqsrtds58pk4a9dg39g7s1h";
+    }
 
     public static function showLimit($page)
     {
