+++ models/repositories/BasketRepository.php
@@ -1,0 +1,27 @@
+<?php
+namespace app\models\repositories;
+
+
+use app\engine\Db;
+use app\models\entities\Basket;
+use app\models\Repository;
+
+class BasketRepository extends Repository
+{
+    public static function getBasket($session)
+    {
+        $sql = "SELECT p.id id_prod, b.id id_basket, p.name, p.description, p.price, p.img FROM basket b, products p WHERE b.product_id=p.id AND session_id = :session";
+        return Db::getInstance()->queryAll($sql, ['session' => $session]);
+    }
+
+    public function getEntityClass()
+    {
+        return Basket::class;
+    }
+
+    public static function getTableName()
+    {
+        return "basket";
+    }
+
+}
