+++ models/repositories/ProductsRepository.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace app\models\repositories;
+
+use app\models\Repository;
+use app\models\entities\Products;
+
+
+class ProductsRepository extends Repository
+{
+    public function getEntityClass()
+    {
+        return Products::class;
+    }
+
+    public static function getTableName()
+    {
+        return "products";
+    }
+}
